move repository interfaces to domain layer, add contracts library
